Check point.
