--- packages/ckeditor5-html-embed/tests/utils.js
@@ -1,7 +1,0 @@
-/**
- * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-
-describe( 'utils', () => {
-} );
